refactor: split sync domain modules
