+++ tests/test-sync-logical-domains.test.ts
@@ -19,10 +19,11 @@ describe("logical cloud sync helpers", () => {
 
   test("maps physical domains into logical groups", () => {
     expect(getLogicalCloudSyncDomainForPhysical("custom-wallpapers")).toBe(
-      "files"
+      "settings"
     );
     expect(getLogicalCloudSyncDomainForPhysical("files-images")).toBe("files");
     expect(getLogicalCloudSyncDomainPhysicalParts("settings")).toEqual([
+      "custom-wallpapers",
       "settings",
     ]);
     expect(getLogicalCloudSyncDomainPhysicalParts("files")).toEqual([
@@ -30,7 +31,6 @@ describe("logical cloud sync helpers", () => {
       "files-trash",
       "files-applets",
       "files-metadata",
-      "custom-wallpapers",
     ]);
   });
 
@@ -80,21 +80,21 @@ describe("logical cloud sync helpers", () => {
     const aggregated = aggregateLogicalCloudSyncMetadata(metadata);
 
     expect(aggregated.settings).toMatchObject({
-      updatedAt: "2026-03-15T10:00:00.000Z",
-      createdAt: "2026-03-15T09:00:00.000Z",
-      totalSize: 100,
+      updatedAt: "2026-03-15T10:05:00.000Z",
+      createdAt: "2026-03-15T08:00:00.000Z",
+      totalSize: 350,
     });
     expect(Object.keys(aggregated.settings?.parts || {}).sort()).toEqual([
+      "custom-wallpapers",
       "settings",
     ]);
 
     expect(aggregated.files).toMatchObject({
-      updatedAt: "2026-03-15T10:05:00.000Z",
-      createdAt: "2026-03-15T08:00:00.000Z",
-      totalSize: 830,
+      updatedAt: "2026-03-15T09:45:00.000Z",
+      createdAt: "2026-03-15T09:00:00.000Z",
+      totalSize: 580,
     });
     expect(Object.keys(aggregated.files?.parts || {}).sort()).toEqual([
-      "custom-wallpapers",
       "files-images",
       "files-metadata",
     ]);
